agripulse/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pandas>=2.0.0
python-multipart==0.0.9
google-generativeai==0.8.3
pydantic==2.9.2
pytest==8.3.3
httpx==0.27.2
python-dotenv==1.0.1
